requirements.txt
# Policy Manual Generation System - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies for document generation
python-docx>=1.1.0        # Word document creation and formatting

# HTTP client for LM Studio communication
requests>=2.31.0          # REST API calls to local LM Studio

# Configuration management
PyYAML>=6.0               # YAML configuration files

# GUI framework (usually included with Python)
# tkinter - Built-in GUI library for Python

# Optional: Enhanced GUI components
# Pillow>=10.0.0           # Image processing for GUI icons
# ttkthemes>=3.2.2         # Additional GUI themes

# Development and testing (optional)
# pytest>=7.4.0           # Unit testing framework
# black>=23.7.0           # Code formatting
# flake8>=6.0.0           # Code linting